Refatora posicionamento de objetos

+++ ce_bitmap.cpp
@@ -47,8 +47,8 @@ unsigned int CEBitmap::getHeight() {
 
 CERenderWindow* CEBitmap::getRenderWindow() {
     CERenderWindow *window = this->getDefaultRenderWindow();
-    CEPoint *start = new CEPoint((int) this->posX, (int) this->posY);
-    CEPoint *end = new CEPoint((int) (this->posX + this->getWidth()), (int) (this->posY + this->getHeight()));
+    CEPoint *start = new CEPoint(this->posX, this->posY);
+    CEPoint *end = new CEPoint(this->posX + (int) this->getWidth(), this->posY + (int) this->getHeight());
     window->setPoints(start, end);
     delete start;
     delete end;
@@ -70,11 +70,11 @@ bool CEBitmap::renderToCanvas(CECanvas *canvas, CERenderWindow *window, CERender
     int endX = window->bottomRight->x < subWindow->bottomRight->x ? window->bottomRight->x : subWindow->bottomRight->x;
     int endY = window->bottomRight->y < subWindow->bottomRight->y ? window->bottomRight->y : subWindow->bottomRight->y;
 
-    unsigned int cursorY = startY;
+    int cursorY = startY;
     unsigned int internalCursorY = 0;
     bool allPixelRendered = true;
     while(cursorY < endY) {
-        unsigned int cursorX = startX;
+        int cursorX = startX;
         unsigned int internalCursorX = 0;
         while (cursorX < endX)
         {
