+++ ce_circle.cpp
@@ -21,8 +21,8 @@ unsigned int CECircle::getRadius()
 CERenderWindow *CECircle::getRenderWindow()
 {
     CERenderWindow *window = this->getDefaultRenderWindow();
-    CEPoint *start = new CEPoint((int)this->posX, (int)this->posY);
-    CEPoint *end = new CEPoint((int)(this->posX + 2 * this->radius), (int)(this->posY + 2 * this->radius));
+    CEPoint *start = new CEPoint(this->posX, this->posY);
+    CEPoint *end = new CEPoint(this->posX + 2 * (int)this->radius, this->posY + 2 * (int)this->radius);
     window->setPoints(start, end);
     delete start;
     delete end;
